renamed tests.common to tests.helpers
