+++ tests/test_area_state.py
@@ -26,7 +26,7 @@ from custom_components.magic_areas.const import (
     AreaStates,
 )
 
-from tests.common import (
+from tests.helpers import (
     assert_in_attribute,
     assert_state,
     get_basic_config_entry_data,
