+++ tests/test_ble_tracker_monitor.py
@@ -20,7 +20,7 @@ from custom_components.magic_areas.const import (
     DOMAIN,
 )
 
-from tests.common import (
+from tests.helpers import (
     assert_in_attribute,
     assert_state,
     get_basic_config_entry_data,
