+++ tests/test_cover.py
@@ -19,7 +19,7 @@ from custom_components.magic_areas.const import (
     DOMAIN,
 )
 
-from tests.common import (
+from tests.helpers import (
     get_basic_config_entry_data,
     init_integration,
     setup_mock_entities,
