+++ tests/test_fan.py
@@ -35,7 +35,7 @@ from custom_components.magic_areas.const import (
     AreaStates,
 )
 
-from tests.common import (
+from tests.helpers import (
     assert_in_attribute,
     assert_state,
     get_basic_config_entry_data,
