+++ tests/test_meta_aggregates.py
@@ -6,7 +6,7 @@ from homeassistant.components.binary_sensor import DOMAIN as BINARY_SENSOR_DOMAI
 from homeassistant.const import STATE_OFF, STATE_ON
 from homeassistant.core import HomeAssistant
 
-from tests.common import assert_state
+from tests.helpers import assert_state
 from tests.const import MockAreaIds
 from tests.mocks import MockBinarySensor
